pyproject.toml
[tool.poetry]
name = "sns_backend"
version = "0.1.0"
description = "사용자는 본 서비스에 접속하여, 게시물을 업로드 하거나 다른 사람의 게시물을 확인하고, 좋아요를 누를 수 있습니다."
authors = ["Yong Su Jeong <knarrblue@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.6"
djangorestframework = "^3.13.1"
python-dotenv = "^0.20.0"
djangorestframework-simplejwt = "^5.2.0"
PyJWT = "^2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
